src/main/java/Excel.java
import java.io.File;
import java.io.IOException;
import java.util.ArrayList;
import java.util.List;
import org.apache.poi.openxml4j.exceptions.InvalidFormatException;
import org.apache.poi.ss.usermodel.Cell;
import org.apache.poi.ss.usermodel.Row;
import org.apache.poi.ss.usermodel.Sheet;
import org.apache.poi.ss.usermodel.Workbook;
import org.apache.poi.xssf.usermodel.XSSFWorkbook;

public class Excel {
    private final Workbook workbook;

    public Excel(File file) {
        this(workbook(file));
    }

    public Excel(Workbook workbook) {
        this.workbook = workbook;
    }

    public List<List<String>> dataset() {
        Sheet sheet = workbook.getSheetAt(0);
        List<List<String>> dataset = new ArrayList<>();
        for (int i = 0; i < sheet.getPhysicalNumberOfRows(); i++) {
            Row row = sheet.getRow(i);
            List<String> rows = new ArrayList<>();
            for (int j = 0; j < row.getPhysicalNumberOfCells(); j++) {
                Cell cell = row.getCell(j);
                rows.add(cell.toString());
            }
            dataset.add(rows);
        }
        return dataset;
    }

    private static Workbook workbook(File file) {
        try {
            return new XSSFWorkbook(file);
        } catch (IOException e) {
            throw new RuntimeException(e);
        } catch (InvalidFormatException e) {
            throw new RuntimeException(e);
        }
    }
}
